Update links to be repo-relative
Updates several links to be relative to the current repo. This enables
forks of this repo to be less dependent on the upstream repo (e.g.
forked, custom images), and importantly makes the links to the forked
repo's lists appropriate (i.e. not pointing at the upstream repo's
files!).

Notably, this hard-codes the 'master' branch, which is not ideal (if
you're browsing a non-master branch, the link will still point to
master), but is no worse than the previous implementation.

Signed-off-by: Chris Ring <chris@ringthis.com>