Add a whitelist file for web developers that use ads
Adds a web_dev.txt list with domains that web developers may want to
occasiontally enable while working on sites that use ads.

Use case is to create a group that only the dev PC is in, and
enable/disable this list during development, when they want to see/test
ad results.

Signed-off-by: Chris Ring <chris@ringthis.com>